MMDevice: Move RegisteredDeviceCollection
Move to an internal (rather than anonymous) namespace in a separate
header, so that it can be tested and reused.
