+++ MMDevice/MMDevice-StaticRuntime.vcxproj.filters
@@ -61,5 +61,8 @@
     <ClInclude Include="Property.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="RegisteredDeviceCollection.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
